Created user property logic
Now user can add a new property only by the Compound.add_property method. Any attempt to add an attribute either to a Compound instance or a UserProperties instance within a Compound object will be rejected.

Setting an already existing property through dot syntax instead is accepted. This should be prevented, though, for the Compound instance attributes - as it might lead to overriding built-in properties such as vendors, chemical_ids and ghs_references. Perhaps this could be done with a @property decorator.